api_server/requirements.unexpanded.txt
connexion==2.7.0
defusedxml==0.6.0
firebase-admin==4.5.1
Flask==1.1.2
Flask-AuditLog==1.0
Flask-Cors==3.0.10
Flask-No-Cache==1.0
Flask-SSLify==0.1.5
Flask-Testing==0.8.1
google-cloud==0.34.0
google-cloud-core==1.5.0
google-cloud-datastore==2.1.0
google-cloud-firestore==2.0.2
google-cloud-logging==2.1.1
google-cloud-secret-manager==2.1.0
google-cloud-storage==1.35.0
gunicorn==20.0.4
jwkaas==0.7
numpy==1.19.5
pandas==1.2.1
pyOpenSSL==20.0.1
PyPDF2==1.26.0
swagger-ui-bundle==0.0.8
Unidecode==1.1.2
Werkzeug==1.0.1
